Move MuJoCo-related files to simulation/mujocoDynamics
